backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
sentence-transformers==3.3.1
scikit-learn==1.5.2
numpy==1.26.4
torch==2.5.1
transformers==4.46.3
python-dotenv==1.0.0